+++ cmd/subcommands/command.go
@@ -1,9 +1,10 @@
 package cmd
 
 import (
-	"fmt"
-	"github.com/ethereum/go-ethereum/rpc"
+	ethereum_rpc "github.com/ethereum/go-ethereum/rpc"
+	"github.com/harmony-one/go-sdk/pkg/common"
 	"github.com/harmony-one/go-sdk/pkg/console"
+	"github.com/harmony-one/go-sdk/pkg/rpc"
 	"github.com/mitchellh/go-homedir"
 	"github.com/spf13/cobra"
 	"log"
@@ -12,26 +13,24 @@ import (
 )
 
 func init() {
-	shardNum := 0
 	net := "mainnet"
 
 	cmdCommand := &cobra.Command{
 		Use:   "command",
 		Short: "Start an interactive JavaScript environment (connect to node)",
 		RunE: func(cmd *cobra.Command, args []string) error {
-			return openConsole(net, shardNum)
+			return openConsole(net)
 		},
 	}
 
-	cmdCommand.Flags().IntVar(&shardNum, "shard", 0, "used shard(default: 0)")
-	cmdCommand.Flags().StringVar(&net, "net", "mainnet", "used net(default: mainnet, choose: mainnet or testnet)")
+	cmdCommand.Flags().StringVar(&net, "net", "mainnet", "used net(default: mainnet, eg: mainnet, testnet ...)")
 
 	RootCmd.AddCommand(cmdCommand)
 }
 
 func checkAndMakeDirIfNeeded() string {
 	userDir, _ := homedir.Dir()
-	hmyCLIDir := path.Join(userDir, ".hmy_cli", "command")
+	hmyCLIDir := path.Join(userDir, common.DefaultConfigDirName, common.DefaultCommandAliasesDirName)
 	if _, err := os.Stat(hmyCLIDir); os.IsNotExist(err) {
 		// Double check with Leo what is right file persmission
 		os.Mkdir(hmyCLIDir, 0700)
@@ -40,37 +39,34 @@ func checkAndMakeDirIfNeeded() string {
 	return hmyCLIDir
 }
 
-// remoteConsole will connect to a remote geth instance, attaching a JavaScript
+// remoteConsole will connect to a remote node instance, attaching a JavaScript
 // console to it.
-func openConsole(net string, shard int) error {
-	var netMap = map[string]string{
-		"mainnet-0": "https://api.harmony.one",
-		"mainnet-1": "https://s1.api.harmony.one",
-		"mainnet-2": "https://s2.api.harmony.one",
-		"mainnet-3": "https://s3.api.harmony.one",
-		"testnet-0": "https://api.s0.b.hmny.io",
-		"testnet-1": "https://api.s1.b.hmny.io",
-		"testnet-2": "https://api.s2.b.hmny.io",
-		"testnet-3": "https://api.s3.b.hmny.io",
-	}
-
-	endpoint := ""
-	if findedEndpoint, ok := netMap[fmt.Sprintf("%s-%d", net, shard)]; ok {
-		endpoint = findedEndpoint
-	} else {
-		log.Fatalf("Unknown network `%s` or shardId `%d`", net, shard)
-	}
-
-	client, err := rpc.Dial(endpoint)
+func openConsole(net string) error {
+	client, err := ethereum_rpc.Dial(node)
 	if err != nil {
-		log.Fatalf("Unable to attach to remote geth: %v", err)
+		log.Fatalf("Unable to attach to remote node: %v", err)
 	}
+
+	// check net type
+	_, err = common.StringToChainID(net)
+	if err != nil {
+		return err
+	}
+
+	// get shard id
+	nodeRPCReply, err := rpc.Request(rpc.Method.GetShardID, node, []interface{}{})
+	if err != nil {
+		return err
+	}
+
+	shard := int(nodeRPCReply["result"].(float64))
+
 	config := console.Config{
 		DataDir: checkAndMakeDirIfNeeded(),
 		DocRoot: ".",
 		Client:  client,
 		Preload: nil,
-		NodeUrl: endpoint,
+		NodeUrl: node,
 		ShardId: shard,
 		Net:     net,
 	}
